+++ STL_Extension/include/CGAL/Handle.h
@@ -39,13 +39,15 @@ class Handle
     
     typedef std::ptrdiff_t Id_type ;
     
-    Handle()
+    Handle() noexcept
 	: PTR(static_cast<Rep*>(0)) {}
 
-    Handle(const Handle& x)
+    // FIXME: if the precondition throws in a noexcept function, the program terminates
+    Handle(const Handle& x) noexcept
     {
       CGAL_precondition( x.PTR != static_cast<Rep*>(0) );
       PTR = x.PTR;
+      CGAL_assume (PTR->count > 0);
       PTR->count++;
     }
 
@@ -56,7 +58,7 @@ class Handle
     }
 
     Handle&
-    operator=(const Handle& x)
+    operator=(const Handle& x) noexcept
     {
       CGAL_precondition( x.PTR != static_cast<Rep*>(0) );
       x.PTR->count++;
@@ -65,6 +67,8 @@ class Handle
       PTR = x.PTR;
       return *this;
     }
+
+    friend void swap(Handle& a, Handle& b) noexcept { std::swap(a.PTR, b.PTR); }
 
     void reset()
     {
@@ -77,11 +81,11 @@ class Handle
     }
 
     int
-    refs()  const { return PTR->count; }
+    refs()  const noexcept { return PTR->count; }
 
-    Id_type id() const { return PTR - static_cast<Rep*>(0); }
+    Id_type id() const noexcept { return PTR - static_cast<Rep*>(0); }
 
-    bool identical(const Handle& h) const { return PTR == h.PTR; }
+    bool identical(const Handle& h) const noexcept { return PTR == h.PTR; }
 
   protected:
     Rep* PTR;
@@ -89,7 +93,7 @@ class Handle
 
 //inline Handle::Id_type id(const Handle& x) { return x.id() ; }
 
-inline bool identical(const Handle &h1, const Handle &h2) { return h1.identical(h2); }
+inline bool identical(const Handle &h1, const Handle &h2) noexcept { return h1.identical(h2); }
 
 } //namespace CGAL